Extract licencas tab into dedicated screen
